Confirm window and fixed item window

+++ app/src/main/java/com/evolgames/entities/usage/Projectile.java
@@ -49,7 +49,7 @@ public class Projectile extends Use implements Penetrating {
             Body overlappedEntityBody = overlappedEntity.getBody();
             worldFacade.freeze(overlappedEntity);
             worldFacade.mergeEntities(overlappedEntity, penetrator, normal.cpy().mul(-actualAdvance), point.cpy());
-            worldFacade.applyPointImpact(obtain(point), (float) (consumedImpulse * massFraction), overlappedEntity);
+            worldFacade.applyPointImpact(obtain(point), consumedImpulse * massFraction, overlappedEntity);
             Invoker.addCustomCommand(overlappedEntity, () -> overlappedEntityBody.applyLinearImpulse(normal.cpy().mul((float) (0.01f * Math.sqrt(consumedImpulse) * massFraction)), point));
         }
     }
